Deal with potential edge case in day 10 solution
Should really test for this but it might be hard to find an example.
The new version will take slightly longer to run but old algorithm was
flawed.

Description of the bug:

After 500 seconds the message was closer to being formed than after 400.
After 600 seconds we get a worse result.
We go back to 500 and start counting from there, but find 510, 520, etc
are worse too. The best score was between 400 and 500.