Solve day 21 part 2
It's not very sophisticated but it works. it doesn't exactly account for
any edge cases (it assumes that the first repeat means we've found all
unique values that will halt the program, which I haven't actually
confirmed is true yet).

There's also a lot of optimisation that could be done.